Changed test folder structure in preparation for ci/cd pipeline setup
